backend/requirements.txt
fastapi
uvicorn[standard]
motor
pydantic
python-dotenv
cloudinary
python-multipart
